Add alternative implementation for determining groups a user belongs to.
This implementation uses the non-POSIX function getgrouplist.
This is available on at least BSD and Linux. The advantage of using
getgrouplist is that it works better with Winbind NSS and nested ADS
groups. The standard getgrnam and getgrgid functions do not pick up a user
that is a member of an ADS group via membership of a nested group.

This alternative implementation must currently be manually selected by
specifying -DHAVE_GETGROUPLIST=1 on the apxs command line.